+++ packages/apply-release-plan/src/index.ts
@@ -93,11 +93,12 @@ export default async function applyReleasePlan(
 
   // iterate over releases updating packages
   let finalisedRelease = releaseWithChangelogs.map(release => {
-    return versionPackage(
-      release,
-      versionsToUpdate,
-      config.updateInternalDependencies
-    );
+    return versionPackage(release, versionsToUpdate, {
+      updateInternalDependencies: config.updateInternalDependencies,
+      onlyUpdatePeerDependentsWhenOutOfRange:
+        config.___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH
+          .onlyUpdatePeerDependentsWhenOutOfRange
+    });
   });
 
   let prettierConfig = await prettier.resolveConfig(cwd);
@@ -134,8 +135,19 @@ export default async function applyReleasePlan(
         let changesetPath = path.resolve(changesetFolder, `${changeset.id}.md`);
         let changesetFolderPath = path.resolve(changesetFolder, changeset.id);
         if (await fs.pathExists(changesetPath)) {
-          touchedFiles.push(changesetPath);
-          await fs.remove(changesetPath);
+          // DO NOT remove changeset for ignored packages
+          // Mixed changeset that contains both ignored packages and not ignored packages are disallowed
+          // At this point, we know there is no such changeset, because otherwise the program would've already failed,
+          // so we just check if any ignored package exists in this changeset, and only remove it if none exists
+          // Ignored list is added in v2, so we don't need to do it for v1 changesets
+          if (
+            !changeset.releases.find(release =>
+              config.ignore.includes(release.name)
+            )
+          ) {
+            touchedFiles.push(changesetPath);
+            await fs.remove(changesetPath);
+          }
           // TO REMOVE LOGIC - this works to remove v1 changesets. We should be removed in the future
         } else if (await fs.pathExists(changesetFolderPath)) {
           touchedFiles.push(changesetFolderPath);
@@ -210,7 +222,12 @@ async function getNewChangelogEntry(
         moddedChangesets,
         getChangelogFuncs,
         changelogOpts,
-        config.updateInternalDependencies
+        {
+          updateInternalDependencies: config.updateInternalDependencies,
+          onlyUpdatePeerDependentsWhenOutOfRange:
+            config.___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH
+              .onlyUpdatePeerDependentsWhenOutOfRange
+        }
       );
 
       return {