+++ packages/apply-release-plan/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @changesets/apply-release-plan
 
+## 4.0.0
+
+### Major Changes
+
+- [`addd725`](https://github.com/atlassian/changesets/commit/addd7256d9251d999251a7c16c0a0b068d557b5d) [#383](https://github.com/atlassian/changesets/pull/383) Thanks [@Feiyang1](https://github.com/Feiyang1)! - Added an experimental flag `onlyUpdatePeerDependentsWhenOutOfRange`. When set to `true`, we only bump peer dependents when peerDependencies are leaving range.
+
+### Minor Changes
+
+- [`9dcc364`](https://github.com/atlassian/changesets/commit/9dcc364bf19e48f8f2824ebaf967d9ef41b6fc04) [#371](https://github.com/atlassian/changesets/pull/371) Thanks [@Feiyang1](https://github.com/Feiyang1)! - Added support for ignoring packages in the `version` command. The version of ignored packages will not be bumped, but their dependencies will still be bumped normally. This is useful when you have private packages, e.g. packages under development. It allows you to make releases for the public packages without changing the version of your private packages. To use the feature, you can define the `ignore` array in the config file with the name of the packages:
+
+  ```
+  {
+    ...
+    "ignore": ["pkg-a", "pkg-b"]
+    ...
+  }
+  ```
+
+  or you can pass the package names to the `--ignore` flag when using cli:
+
+  ```
+  yarn changeset version --ignore pkg-a --ignore --pkg-b
+  ```
+
+### Patch Changes
+
+- Updated dependencies [[`addd725`](https://github.com/atlassian/changesets/commit/addd7256d9251d999251a7c16c0a0b068d557b5d), [`9dcc364`](https://github.com/atlassian/changesets/commit/9dcc364bf19e48f8f2824ebaf967d9ef41b6fc04)]:
+  - @changesets/config@1.2.0
+  - @changesets/types@3.1.0
+
 ## 3.1.0
 
 ### Minor Changes